Update .env.server.example
The docs at https://docs.opensaas.sh/guides/email-sending/ claim that:

"In the .env.server file, we’ve set the SEND_EMAILS_IN_DEVELOPMENT env variable to true. This means that emails will be sent in development mode.

This is useful for testing, but you can turn it off by setting it to false, and the emails will be logged to the console instead.".

However, this isn't in code.